requirements.txt
aspose-pdf==24.1.0
certifi==2024.2.2
python-dateutil==2.8.2
six==1.16.0
urllib3==2.2.0
